change semantics of a2acl_fromfile
Simplify return value to either 0 on succes, or -1 on error. Furthermore, pass
back information about both the number of newly imported rules, like before and
the number of total rules. This last piece of information is important when
caches are re-used with the LMDB backend. Simply knowing that zero new rules are
imported is not enough for a caller to be sure the database is not empty. Now it
can look at the total number of rules as well.