refactor a2id_tostr into a more idiomatic and useful API
This new function has a simpler api and is easier to use in practice
thanks to the fact that it returns the needed space for the string, like
snprintf. With the old API it was not possible to know what went wrong
or how much space is needed to make it right.